Add files via upload
This is my import of my python script to build a bibliography from a word (.docx or .doc) file containing cites such as "(Authors, date)" or "Authors (date)" and a bibtex file (.bib). The script requires a path to a .docx file and a .bib (asked as inputs when running the script).

This is basically an attempt to save time when building a bibliograpghy outisde of LaTeX and without a proper word add-on.